organize the files
