+++ burger-queen-project/src/components/menu/boxOrder.js
@@ -1,10 +1,26 @@
 import styles from "../../styles/boxOrder.module.css";
+import { useState } from "react";
 
-export function BoxOrder () {    
+export function BoxOrder() {
+    const [order, setOrder] = useState([{
+        products: [],
+        total: 0,
+        customer: { name: '' }
+    }]);
+    const handleOrder = (e) => {
+        const { nameProduct, value } = e.target.value;
+        setOrder((preOrder) => ({
+            ...preOrder,
+            products: [preOrder.products],
+        }))
+    };
     return (
         <div className={styles.burgerBox}>
-                <label className={styles.customer}>Cliente: <input type="text" className={styles.customerName}></input></label>
+            <label className={styles.customer}>Cliente: <input type="text" className={styles.customerName}></input></label>
+            <div>
+                <p >Tabla de ordenes</p>
+            </div>
         </div>
-        
+
     )
-  }
+}
