+++ burger-queen-project/src/pages/_document.jsx
@@ -1,10 +1,14 @@
-import { HeaderMenu } from '@/components/menu/headerMenu'
+import { HeaderMenu } from '../components/menu/headerMenu' 
 import { Html, Head, Main, NextScript } from 'next/document'
+// export const metadata = {
+//   title: 'Burger Queen',
+//   description: 'generate by create NextApp',
+// }
 
 export default function Document() {
   return (
     <Html lang="en">
-      <Head />
+      <Head metadata />
       <HeaderMenu />
       <body>
         <Main />
