+++ burger-queen-project/src/pages/food.jsx
@@ -1,15 +1,12 @@
 import { HeaderMenu } from "../components/menu/headerMenu";
 import { NavigationMenu } from "../components/menu/navigationMenu";
-import { BoXOrder } from "../components/menu/boxOrder";
 // import { BoxCategoriesMenuFood } from "../components/menu/boxCategoriesMenuFood";
-import { BoxProcess } from "../components/menu/boxProcess";
-import { AllFood } from "@/components/menu/all-food";
+import { ViewAllFood } from "../components/menu/viewAllFood";
 
 export default function MenuFoodPage() {
     return (<div>
         <NavigationMenu />
-        <AllFood title="Desayuno"/>
-        <BoxProcess />
+        <ViewAllFood title="Almuerzo"/>
     </div>
     )
 }
