+++ burger-queen-project/src/pages/index.jsx
@@ -2,6 +2,7 @@ import Image from 'next/image';
 import { Inter } from 'next/font/google';
 import { useRouter } from 'next/router';
 
+
 const inter = Inter({ subsets: ['latin'] })
 
 export default function Home() {
