+++ burger-queen-project/src/pages/menu.jsx
@@ -1,16 +1,16 @@
 import { HeaderMenu } from "../components/menu/headerMenu";
 import { NavigationMenu } from "../components/menu/navigationMenu";
 import { BoXBurgerDance } from "../components/menu/boxBurgerDance";
-import { BoXOrder } from "../components/menu/boxOrder";
-import { BoxProcess } from "../components/menu/boxProcess";
+import { BoxOrder } from "../components/menu/boxOrder";
+import { BtnProcessOrder } from "../components/menu/btnProcessOrder";
 // import { AllFood } from "../components/menu/all-food";
 
 export default function MenuPage() {
     return (<div>
         <NavigationMenu />
         <BoXBurgerDance />
-        <BoXOrder />
-        <BoxProcess />
+        <BoxOrder />
+        <BtnProcessOrder />
     </div>
     )
 }
